refactor(configurations): Move Entities to configuration package
Refactoring the code to be in the "configuration" package, and delete the deprecated Configuration class.
