+++ src/main/kotlin/com/expediagroup/sdk/core/config/package-info.java
@@ -1,0 +1,3 @@
+package com.expediagroup.sdk.core.config;
+
+// TODO: WiP - this package will be moved and merged with the existing Configuration package.
