--- src/test/kotlin/com/expediagroup/sdk/core/config/ConfigurationTest.kt
@@ -1,35 +1,0 @@
-/*
- * Copyright (C) 2022 Expedia, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.expediagroup.sdk.core.config
-
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Assertions.assertNotNull
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Test
-
-class ConfigurationTest {
-    @Test
-    @DisplayName("defaults with overloaded constructor")
-    fun testConfigurationOverload() {
-        val credentials = Credentials("id", "secret")
-        val configuration = Configuration(credentials)
-
-        assertThat(configuration.credentials).isSameAs(credentials)
-        assertNotNull(configuration.network)
-        assertNotNull(configuration.logging)
-        assertThat(configuration.logging.level).isEqualTo(LoggingConfiguration.LogLevel.WARNING)
-    }
-}
